+++ fabric/src/main/java/net/satisfy/meadow/fabric/MeadowFabric.java
@@ -4,11 +4,11 @@ import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.serializer.GsonConfigSerializer;
 import net.fabricmc.api.ModInitializer;
 import net.satisfy.meadow.Meadow;
+import net.satisfy.meadow.core.registry.CompostableRegistry;
 import net.satisfy.meadow.fabric.core.config.MeadowFabricConfig;
 import net.satisfy.meadow.fabric.core.villager.FabricVillager;
 import net.satisfy.meadow.fabric.core.world.FeatureModification;
 import net.satisfy.meadow.fabric.core.world.MeadowBiomeModification;
-import net.satisfy.meadow.core.registry.CompostableRegistry;
 
 
 public class MeadowFabric implements ModInitializer {