+++ fabric/src/main/java/net/satisfy/meadow/fabric/core/world/MeadowBiomeModification.java
@@ -1,6 +1,9 @@
 package net.satisfy.meadow.fabric.core.world;
 
-import net.fabricmc.fabric.api.biome.v1.*;
+import net.fabricmc.fabric.api.biome.v1.BiomeModifications;
+import net.fabricmc.fabric.api.biome.v1.BiomeSelectionContext;
+import net.fabricmc.fabric.api.biome.v1.BiomeSelectors;
+import net.fabricmc.fabric.api.biome.v1.ModificationPhase;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.SpawnPlacementTypes;